DateTimeController: Fix integer overflow

+++ src/components/datetime/DateTimeController.h
@@ -67,7 +67,7 @@ namespace Pinetime {
         return static_cast<Days>(daysSinceSunday);
       }
 
-      uint8_t DayOfYear() const {
+      int DayOfYear() const {
         return localTime.tm_yday + 1;
       }
